qrcode-generator/pyproject.toml
[tool.poetry]
name = "qrcode-generator"
version = "0.1.0"
description = ""
authors = ["thinpyai <thinpyaiwin@gmail.com>"]
readme = "README.md"
packages = [
    { include = "*", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
qrcode = "^7.4.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
autopep8 = "^2.0.4"
pylint = "^3.0.3"
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
coverage = "^7.4.0"
freezegun = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
